third_party/WebKit/LayoutTests/external/wpt/web-animations/animation-model/combining-effects/effect-composition-expected.txt

This is a testharness.js-based test.
FAIL accumulate onto the base valueassert_equals: Animated margin-left style at 50% expected "15px" but got "5px"
FAIL accumulate onto an underlying animation valueassert_equals: Animated style at 50% expected "20px" but got "5px"
FAIL Composite when mixing accumulate and replaceassert_equals: Animated style at 50% expected "25px" but got "20px"
FAIL accumulate specified on a keyframe overrides the composite mode of the effectassert_equals: Animated style at 50% expected "20px" but got "15px"
FAIL unspecified composite mode on a keyframe is overriden by setting accumulate of the effectassert_equals: Animated style at 50% expected "20px" but got "15px"
PASS add onto the base value
PASS add onto an underlying animation value
PASS Composite when mixing add and replace
FAIL add specified on a keyframe overrides the composite mode of the effectassert_equals: Animated style at 50% expected "20px" but got "15px"
FAIL unspecified composite mode on a keyframe is overriden by setting add of the effectassert_equals: Animated style at 50% expected "20px" but got "15px"
Harness: the test ran to completion.
